[ENC-307] Add support for resolving generic names
This commit adds support to the names package so we can resolve types which use
more than 1 generic type (i.e. `X[A, B]`) as this uses a new ast type which we
previously didn't support.

To make this backwards compitable, I've refactored the names package and introduced
the concept of "language levels" which allow us to add logic per-language level, which
we can use the built in language level tags to enable/disable those source files.

Meaning this change will still compile against Go 1.16 even though it's using a type
only avaliable in the Go 1.18 version of AST package.