[ENC-307] Generic Response Types
This commit gets generic response types working under Go 1.18 and
allows the system to return a response struct with a generic field
within it.

This commit also adds the ability for the encore parser to understand
and obey `//go:build` tags, which helps with things like ARCH, OS,
language level or tooling tags filters on files - preventing us trying
to parse a file we can't or shouldn't.