[ENC-310] Update the dashapp to support generics
This commit updates the `dashapp` so it works with generic types.
While I was here I removed some duplication of code in the JSON
default parameters for requests. As part of unifiying the dupliated
code I've moved the SchemaView over to using `CodeMirror` as well
so we get syntax highlighting in the API docs.

I've also updated the response renderers in both `RPCCaller` and
the trace overview to also render the payloads with syntax highlighting.