v2: fix codegen generated imports
When we injected an import during the rewriting of a file we would
add the import right before the first import declaration.

This approach had the generally-innocent side effect of separating the
import declaration from any comment attached to the import. This was
almost never a problem.

But when that import happened to be a cgo import it caused confusing compilation
errors, since the `import "C"` comments are semantically important.

Fix this by always adding the imports right after the package name.

Thanks Patryk Siemiński for the report.