runtimes/{core,js}: refactor schema encode/decode
Introduce a pair of types, `PValue` and `PValues`. The types
allow us to introduce dedicated semantic types that go beyond
plain JSON types.

As a result, since the wire protocol is still JSON, decoding JSON
into these types requires an API schema (so that we know e.g.
whether a string is "just a string" or should be a datetime.

This commit introduces the concept and refactors the runtimes
to use it. The next commit adds support for dates.